fix: don't perform splash attack if aoo or splash target dead
This fixes an issue where killing a fleeing enemy upon attack of opportunity would cause the game to freeze if there was an entity next to that target who got hit by the splash attack and died. The issue occurred if that secondary entity was next in line in the turn order after the primary target. This is because of how the game handles the timing of the attack and the initialization of the next turn.

This fixes the situation e.g. with Lindwurms where killing the head causes a crash because the tail is also dead.