refactor: reduce hit chance bonuses of slash, thrust, gash
This is to account for the generally longer Reach of spears and swords compared to other one-handed weapons.

This is a temporary solution and includes the most important skills that need adjustment. Ideally we'd have a more streamlined way to adjust both hitchance and tooltip of all skills. That requires a lot of work as the vanilla skills especially 2 tile ones like Impale etc. are very rigidly coded in terms of their tooltip, bonuses (onAnySkillUsed), mastery checks, and this.m.HitChanceBonus.