fix: use setContainer and onEquip instead of container.equip
Because somehow using equip() causes the "faction_banner" item to throw an error on alternate launches of the game. More weirdly, if it works on a launch then it'll work for a campaign started during that launch. If it was the launch where it errors, then it won't work for a campaign started in that launch. The behavior in the campaign persists for all future launches of the game. Setting container manually and calling onEquip doesn't suffer from this issue. This should be looked into at some point.