fix: use isEngagedInMelee instead of Tactical.isActive
This is useful for nested tooltips of skills being queried during the loot screen, where tactical is active but we don't want the tile of the dummy player to be accessed - otherwise it leads to a crash. The isEngagedInMelee function is an MSU function that makes the actor.isPlacedOnMap() check so it should avoid the crash.