refactor: call update on receivers via skill container instead of skills
This improves performance when multiple auras are present on a character. In the original implementation, all of those auras will each trigger an update for the receivers. Whereas, in this implementation only a single update is triggered for receivers as long as they receive at least one aura from this character.