feat: copy dynamic spawns data from dynamic spawns framework
The original data was prepared by Darxo. This commit copies the original data from the framework, so that the implementation of the data can be removed from the framework. For data where we already have a Reforged implementation, that implementation is kept. The IDs are also adjusted to conform to the new convention of unit and unitblock IDs starting with `Unit.` and `UnitBlock.` respectively.

Co-authored-by: Darxo <Darxo@KAISERTOM>