refactor: remove stagger and instead apply stun if attack is successful
This fixes an issue whereby the application of Staggered causes a crash because of the switch entities that happens later. The staggered application pushes the entity back in the turn order which calls `onEntityEntersFirstSlot` in turn_sequence_bar for the attacker, which calls `onTurnResumed` of actor which calls `updateVisibility` where the tile is passed. During entities being switched by the navigator, the tile they are placed in is ID 0 which means they are not placed on map so anything to do with tiles crashes the game.