fix: prefer to fetch perk icon from Const.Perks instead of getIcon
Some vanilla perks have an incorrect icon defined in their skill file (for example vanilla mastery skills use the perk_10.png icon which is the anticipation icon). This leads to issues in tactical tooltip when displaying the perk icon. Fetching the icon from Const.Perks directly ensures that the icon that's used in the perk tree is the one that's fetched.