feat: don't trigger positive morale checks if killer is not allied to me
This is a change from vanilla behavior. In vanilla if an entity dies, then "hostile" actors get a positive morale check irrespective of how that entity died. In Reforged we change this so that the positive morale check is triggered only if the killer is null or belongs to an allied faction.