fix: case in soul link skill where transfer target could be null
An example of the edge case is where a Hexe would charm a target making it a valid link. Then the Hexe takes damage, when the link is still valid. Then it dies causing the Charm to expire and the link to no longer be valid and the transfer target being null - causing the Hexe to bug out and become unkillable.