refactor: improve and cleanup kata step ai
Create a list of possible kata paths up to 2 consecutive usages of Kata Step. For each tile along the path evaluate all targets that could be attacked from that tile using all of my skills. Choose the path along which we would have the best possible target to attack.

Add check for being a bodyguard, in which case prefer the kata path that would keep you close to the VIP.

Improve performance by setting this.m.IsThreaded to true and adding a yield statement between each path evaluation.

Change score and order to those of the vanilla Disengage behavior.