fix: triggering when the character had resisted the morale check
When a non-zero parameter is passed as the first parameter to the checkMorale function it returns false if the morale state didn't change (i.e. the check didn't go through) and returns true if the morale state did change. When 0 is passed then it returns false if the check didn't go through (i.e. the actor resisted the morale check) and returns true if the check passed and the character couldn't resist it. However, in this case the morale state doesn't change. See vanilla usage of this 0 type morale check in charm_skill and horror_skill.