refactor: rebalance the stats of two handed cleavers
- Kriegsmesser: Increase damage and remove additional fatigue cost of attacks.
- Great Khopesh: Increase damage and armor penetration.
- Crypt Cleaver: Increase armor penetration.
- Rusty Warblade: Increase damage, armor damage and shield damage. Change to using great cleave (6 AP) and 6 AP decapitate.

The idea is to make the Crypt Cleaver better in damage than the Two Handed Scimitar because it is a lootable-only weapon and has low durability. The Rusty Warblade was deemed to be too high of damage output with 4 AP attacks so switching it to 6 AP leaves the damage output similar per turn compared to the old setup but in a different way (e.g. fewer bleed inflicted and changed behavior with AP related perks). The Kriegsmesser does less damage than the Two Handed Scimitar but builds less Fatigue and has the versatility of being able to use the sword perks.