fix: resource usage and cost for units with subparties (#373)
Dynamic Spawns Framework 0.3.0 adds the cost of the spawned party into the cost of the unit when subtracting its worth from the top party's resources. So we add a check to not use resources from the top party, and remove the manual addition of the extra cost from the unit with the subparty.