fix: combo updating used skill too early and not applying AP cost bonus
This fixes the issue where you used a skill leading to reduced cost of other skills, but when you use one of the other skills its cost is reset back to what it was before Combo's bonus because the combo used skill updated even before the skill got executed. Therefore, the usage of the other skills still ended up costing AP without the reduction from this perk.