fix: move reload skills to be ordered immediately after shoot skill
The previous method worked in the scenario where we didn't remove these skills during use. However, after reverting back to the vanilla method of adding/removing these skills during their use, it is now necessary to do it in this way. This method assumes that shooting is always the first skill of the weapon.