refactor: add separate functions for preparing and dropping loot
This allows mods to modify the loot that is going to be dropped without having to manually drop the loot they want. This is more compatible and avoids cases where the same original loot is dropped multiple times by several mods because they think they are responsible for dropping the loot.