refactor: don't add damage entries to between the eyes tooltip
Instead use base skill getTooltip - because the user can mouse over the hyperlink of the AOO to see its damage values. This is because some attacks add custom entries to the tooltip and we won't be able to reflect that here otherwise (e.g. any extra damage or other modifiers on an attack).