fix: killer being inconsistent between kill and onDeath functions
Because we were changing the killer in `onDeath` which is called from `kill`, it was causing `kill` to have a different killer than `onDeath`. In this commit's method, we change the killer during `kill` which has the advantage of keeping the killer consistent across both functions and also more clearly shows intent that we only consider player or player animal faction to be the killer if he did at least 50% damage.