refactor: remove changing killer during onDeath (#624)
The purpose of this feature was to make loot drop only if the player and player animals did at least 50% damage to the dying entity, regardless of who the killer is. However, this was accomplished by changing the `_killer` which leads to conflicts with other mods which expect the `_killer` to be the real killer. So, in view of greater compatibility with mods, this feature should be removed until a more stable and more compatible method of accomplishing this can be achieved.