fix: double grip free swap working opposite than intended
The intention was to allow swapping to daggers in the bag to be free, but it was working the opposite way because vanilla recalculates and pays the item action cost AFTER the item action is complete. By that time the dagger is no longer in the bag so `item.isInBag()` returns false.

An ideal solution would be to add LastSlotType to items which holds information about the last slot they were equipped in (in addition to the CurrentSlotType that they have) and that could be used here.

But the change in this commit is intended to provide a middle-ground design while we don't have that tech.