fix: unstoppable not resetting to 0 upon getting stunned or rooted
Because the IsStunned and IsRooted properties are set during `onUpdate` of the effects that cause them, SkillOrder becomes important if we are checking for those properties in `onUpdate`. The stunned_effect may be present but if it updates after this perk then we won't detect the character being stunned. We can't check for `CurrentProperties.IsStunned` either because CurrentProperties are only updated at the end of the entire update cycle and if the stun comes from an effect that updates after this perk, we will still miss it.

However, this will now cause the things done by this perk to happen a bit too late for other skills which may depend on those stats (e.g. increased Initiative). But that seems correct because otherwise that also depends on SkillOrder which seems wrong from a design pov.