refactor: remove tile parameter from onMovementFinished
onMovementFinished was added by vanilla in 1.5.1.4 and that function has no parameter. The one with parameter was added by MSU before the vanilla update. This parameter was not causing any issues because MSU has legacy support for functions with the parameter, but we should still remove it for consistency with vanilla.