fix: only climb up to the first non-native function
Otherwise things scheduled from other classes as a result of a skill execution e.g. inside the onDeath of an entity killed by a skill, would also be considered as part of a skill's scheduled events. An example is the resurrection of zombies which is scheduled for several rounds later using scheduleEvent with TimeUnit.Rounds).

Note: As a result of this change it is now required that any delayed event that should be considered part of a skill's execution must be executed with that skill as the env variable.