feat: allow the AI to evaluate in parallel but rethink if state changed
This should improve AI performance because we allow the AI to evaluate (but not execute) its behaviors while delayed events are scheduled from a used skill. However, if anything important changed in the state during the delayed events e.g. anyone died, or moved, or morale state changed, or Action Points changed, then we force the AI to reevaluate using this new state. This is because any important state change could make the already evaluated behaviors to be no longer valid for execution.