fix: add handling for instant teleport and switchEntities
The teleport function happens immediately in fog of war, so we can't use a check for Count == 0 to trigger the function manually as it leads to one additional call than intended.

The switchEntities vanilla function does not call its callbacks when used in fog of war. This seems like a vanilla bug, so we fix it by calling the callbacks of switchEntities manually from actor.onMovementFinish. In vanilla these callbacks happen after onMovementFinish when not in fog of war so the flow of events should remain unchanged.