feat: invalidate agent when an active skill is removed from its actor
This covers all cases where a delayed effect may cause the skill which had been chosen by a behavior to now be removed from the actor. Also prevents already evaluated behaviors to now no longer be valid for being picked because the skill they were evaluated with is potentially gone. Examples include when a weapon breaks as a result of an attack (the breakage is a delayed event in vanilla).