fix: crash when triggering kingfisher on target killed by flail spinner
The crash occurs because the first hit from the flail uses throw_net due to Kingfisher. This schedules onNetSpawn. Meanwhile the hit also schedules the second attack from Flail Spinner. This second attack happens before onNetSpawn and kills the target. Now onNetSpawn triggers and, due to the entity having died, crashes when trying to access the entity's sprites (in the vanilla code).