feat: decouple ai agent canExecute from scheduled skills
Instead we use Navigator.IsTravelling to stop execution during teleport and switchEntities, and we use Time.hasEventScheduled to stop execution while events are scheduled with VirtualTime i.e. delayed effects of skills.

Also rename __IsStateSaved to __IsExecuting to be a lot clearer about when and where this is used. And add an exception for this in RF_canExecute to allow agent.think to fall through to agent.execute during a behavior's execution.