fix: onNetSpawn crashing with dead actor still considered placed on map
Apparently vanilla still returns true for `isPlacedOnMap` for some time for an actor even after an actor has died. Apparently the ID of the actor also changes in the meantime. Example: I attacked a Peasant in a way that the attack killed him while also executed a Throw Net on his tile. In the `onRemovedFromMap` function of the Peasant he printed his ID as 24110 whereas in the onNetSpawn function he printed it as 1936941420. In both functions `isPlacedOnMap` was true while `isAlive` was false and `isDying` was true.