fix: flaming arrows with weapon master causing bow skills to persist
Because upon unequipping the bow, the Weapon Master perk causes the perks of the bow perk group to be removed. This then triggers the onRemoved of the Flaming Arrows perk which causes the bow (which hasn't been unequipped yet because onUnequip runs before unequipping) to be unequipped and reequipped which re-adds the skills of the bow which go into skill_container.m.SkillsToAdd. Eventually once the onUnequip event is complete, and the bow is fully unequipped, a skill_container update happens which then puts these skills onto the user and they are now hanging without an item. Therefore, instead of unequipping and equipping the mainhand item we instead remove the Flaming Arrows active skill directly upon removal of the Flaming Arrows perk.