fix: inspiring presence setting action points during onUpdate
It is not good to change/set values during onUpdate in principle because it is logically meant to be a place where `_properties` are updated. In any case, there is a vanilla bug in `dislocated_shoulder_injury` where action points are set during `onUpdate` which causes changes to `_properties.ActionPoints` from other skills to not be properly taken into account. That should move to `onAfterUpdate` and therefore it also makes sense that we move it in this perk to `onAfterUpdate`.