feat(chg): rework deep impact
- No longer a proxy skill, instead is a new skill of its own.
- Streamline code to be a hammer only skill.
- Reduce Fatigue Cost from 30 to 25.
- Costs 4 AP with one-handed but 6 AP with two-handed hammers.
- Buff Deep Impact effect to also reduce melee skill, ranged skill, melee defense and ranged defense in addition to damage.

fix: missing space in tooltip

Co-authored-by: Darxo <Darxo@users.noreply.github.com>

fix: set deep impact as weapon skill

fix: only set to two-handed ap cost for 2handed weapons

fix: using flaming arrows id instead of deep impact skill id

refactor: use onItemSet to adjust AP cost on two handed hammers

feat: convert deep impact to a dummy skill that converts a base skill

This solves the problem where it was inheriting from `smite` causing it to always Stagger the target even with a one-handed hammer. The new implementation means that rf_deep_impact_skill is never added directly to a character. Rather, a base skill is added e.g. smite_skill and we "convert" that skill to a deep impact skill by overwriting some of its m table and extending some of its functions. This ensures that one-handed hammers get a deep impact Batter and two-handed ones get a deep impact Smite.

As the ID of this "converted" skill is changed to `actives.rf_deep_impact` that means that any other skills that target that base skill e.g. `actives.smite` to adjust its properties will not apply to the Deep Impact version of it. This is an acceptable compromise and we regard Deep Impact in this case to be a standalone version of that skill that behaves differently.