feat: call onSpawn via hook on Tactical spawnEntity function (#858)
This eliminates all the complicated logic about temporarily blocking onSpawn from triggering before equipping of items etc. It also eliminates the bug whereby certain edge cases were not triggering onSpawn e.g. when an entity was spawned from within a skill but which did not act immediately or was not the result of a resurrection.