fix: swordmaster skills saying requires sword even when sword equipped
The parent skill checks that a sword is equipped via calling isEnabled(). However children may overwrite this function to add custom functionality e.g. additional requirements for the sword. This causes this function to print its own required tooltip which is both redundant and inaccurate as the character may have a sword equipped.