feat: convert deep impact to a dummy skill that converts a base skill
This solves the problem where it was inheriting from `smite` causing it to always Stagger the target even with a one-handed hammer. The new implementation means that rf_deep_impact_skill is never added directly to a character. Rather, a base skill is added e.g. smite_skill and we "convert" that skill to a deep impact skill by overwriting some of its m table and extending some of its functions. This ensures that one-handed hammers get a deep impact Batter and two-handed ones get a deep impact Smite.

As the ID of this "converted" skill is changed to `actives.rf_deep_impact` that means that any other skills that target that base skill e.g. `actives.smite` to adjust its properties will not apply to the Deep Impact version of it. This is an acceptable compromise and we regard Deep Impact in this case to be a standalone version of that skill that behaves differently.