feat: add handling and info for AOOs in tile tooltip on movement preview (#878)
Adds the following information to the tile tooltip during movement preview:
- Fatigue build-up for evading all attacks of opportunity at the starting tile.
- Chance to be hit by attacks of opportunity at the starting tile.
- Chance to be hit by attacks of opportunity at the ending tile (e.g. spearwall).

Adjusts the TacticalNavigator `buildVisualisation` and `getCostForPath` functions to account for the necessary fatigue build-up from dodging the attacks of opportunity at the starting tile. This will prevent the edge case where the game would allow you to queue a movement which will certainly be stopped due to evading AOOs if evading the AOOs pushes your Fatigue to a level where you cannot afford the movement cost of the first tile on your path. This edge case should be considered a vanilla oversight/bug.