fix: crash in tooltip due to rf_reach if skill gets tile during onUpdate
Because during onQueryTooltip in the rf_reach skill we set the container to dummy player's skills and dummy player isn't placed on map. So any tile access will lead to problems, this includes `serpent_racial.onUpdate` as a vanilla example which will cause a crash.

Now we just call the onUpdate for the skill on its existing container, and then force an update on that container afterward.