feat: move damage of some vanilla skills to onAnySkillUsed
This is only important to do for skills which are present on enemies who can have multiple sources of damage modification in `onUpdate` active at the same time e.g. enemies who carry weapons. For unarmed enemies e.g. Unholds, Beasts etc. it is not necessary to do this. A consideration with them is also that they sometimes have a `_zoc` version of their main skill that depends on the damage from the main skill to be added to onUpdate. This *could* be changed by changing the inheritance of those skills to inherit from the main skill, but seems unnecessary to do.