fix: newskilluse check not working for all skills present on actor
Because the logic was present on skill_container, this meant that the function would return true only for the first skill that calls it, and then would return false for all subsequent calls. To work correctly, it needs to return true for each skill the first time that skill calls it, therefore, we should put this on the skills rather than skill_container.