fix: hit chance bonus for skills not applying after modular vanilla fix
Modular Vanilla "fixed" the vanilla issue of `HitChanceBonus` being inaccurate for skills. This fix involves setting `m.HitChanceBonus` of all skills to 0 by default and then setting it to the correct value in the `onAnySkillUsed` function based on the change in `_properties.MeleeSkill`. This allows other skills to additively change `HitChanceBonus` of the used skill too. But this also means that we cannot use `HitChanceBonus` as a field to contain usable configuration for a skill. That must instead be done by a separate member variable.