fix: not expanding LevelXP when XP is added directly
Some mods add XP directly to a character and then call updateLevel. This relevant for some debug mods which may add level. So we expand the LevelXP array just before updateLevel based on the current XP of the character. We also expand it on a call to `getXPForNextLevel` for the edge case where this is called at a time before the LevelXP array has been expanded.