fix: attack scheduling logic for flail spinner and whirling death (#919)
- Trigger Flail Spinner only after a complete execution of a skill. For this, we wait until that skill isn't being executed i.e. it has no schedules. This ensures that Flail Spinner properly behaves as a second attack after everything related to the previous one is complete.
- Queue Whirling Death in onAnySkillExecuted but trigger it only after a complete execution of a skill.

This now results in proper intended behavior even with skills that do multiple attacks per skill use (e.g. vanilla cascade and hail skills for three-headed flail). Now it will properly do: Attack, Whirling Death, Flail Spinner Attack, Whirling Death.

This also removes the need to check in Whirling Death for whether the actor is Spinning Flail before applying the damage multiplier. This keeps both skills completely encapsulated in their own scripts.

Also changes the spinning or whirling to always be conducted via a `scheduleEvent` instead of doing it sequentially when in fog of war. This may have been causing some hidden issues because during onAnySkillExecuted or onAnySkillExecutedFully the skill container IsUpdating is true, and we are not supposed to execute skills.

Reorganize code in both skill files to make it cleaner and more modular.

Incrementing SkillCounter++ is necessary because vanilla has some effects that trigger based on SkillCounter e.g. condition loss on three-headed flail. And in order for these Flail Spinner and Whirling Death to behave properly as "new usages of the skill" we need to increment SkillCounter.