fix: weapon master sometimes not adding perks correctly during combat
Currently Weapon Master does not add perks correctly when switching from
the same weapon to the same weapon. Instead the perks vanish.

That happens, because a recent new rule in Weapon Master preserves the
previous perks until after the switch (so that swap discounts work)
But during the switch, new perks are only added, if those are not
already present.