fix: bone breaker and mauler sometimes not applying their extra injury
The MV_applyInjury function first selects an injury and that happens based on the injury threshold calculation. So, if we want a guaranteed injury we need to reduce the injury threshold to an extremely small value. We don't set it to 0.0 to avoid potential issues with things that don't expect this value to be 0.